Issue #1946: Add check for customer user password

+++ Kernel/Modules/AdminCustomerUser.pm
@@ -371,7 +371,7 @@ sub Run {
                 );
             }
 
-            if ( $CurrentUserData{UserPassword} ne $GetParam{UserPassword} ) {
+            if ( $CurrentUserData{UserPassword} && $CurrentUserData{UserPassword} ne $GetParam{UserPassword} ) {
 
                 $UpdateSuccess = $CustomerUserObject->DeleteOnePreference(
                     Key    => 'UserLastPwChangeTime',
